modified:   src/policy_heads/models/droid_unet_diffusion.py

+++ src/policy_heads/models/droid_unet_diffusion.py
@@ -184,7 +184,7 @@ class ConditionalUnet1D(nn.Module):
 
         dsed = diffusion_step_embed_dim
         diffusion_step_encoder = nn.Sequential(
-            SinusoidalPosEmb(dsed, torch.bfloat16),
+            SinusoidalPosEmb(dsed, torch.float32),
             nn.Linear(dsed, dsed * 4),
             nn.Mish(),
             nn.Linear(dsed * 4, dsed),
